Mockup screen appearance
I made a mockup in Affinity Designer and then implemented all of the static parts. Pretty much all that's left is to swap out image and draw text (for text area and choices) based on current screen.

I'll be using the Raylib bounded text example to get that done.

Right now everything is in garish colors just to make sure everything works. I need to select a color palette and fonts soon and get those implemented.