Start adding screens and side effect functionality
I was going to use XML to generate my game's screens, but I decided with my time limited and my skills lacking, I should just hardcode all the data. It's a work in progress.

The basic loop is: a screen is loaded up, the player can click a choice, which runs a list of side effects (or not) and loads a new screen (in that order). So I need to effectively load a screen's text and choice text in, allow player to click the choices, and ensure they load a new screen and run their side effects correctly.

All very doable.

I also added one more image I forgot I needed.