Modify owl screen and owl loop
The owl loop is very lazy now to save a little work and because I think it's kind of funny. It does commit a game dev sin of not giving the player feedback, so I will open an issue to add a sound effect in case I work on this game more in the future. Maybe I'll get it done before turning the game in.